+++ CHANGELOG.md
@@ -9,3 +9,6 @@ Initial release
 * Removes disabled configuration for admin via payment status
 ### 1.0.3 (2024-07-30)
 * Add User-Agent to Request Transaction 
+### 1.0.4 (2024-10-23)
+* Resolve deprecated function to generate mach pay qr
+* Added support for when MachPay is set to manual mode and order confirmation must be sent from the website